[mod-braille] dotify-formatter: smarter placement of markers leading to more correct page ranges
Page numbers are not incorrectly moved to the next range anymore.

See https://github.com/brailleapps/dotify/issues/150#issuecomment-245373832

There is still a remaining issue which apparently is impossible to fix
without changing OBFL/Dotify. (There seems to be no way, even by using
complex combinations of markers with different classes, special CSS
processing of marker-references, etc.)

Page ranges would be more correct when, in the case of pagenums
coinciding with page breaks, those page numbers would be in only one
page range (the one after the break) rather than in both. With the
current implementation, the last print page of a certain range is always
the same as the starting print page of the next range, regardless of
where the pagenums happen exactly.

The fundamental problem is that page ranges must behave like this
because the current implementation of start-except-last relies on
it (and cannot be easily implemented another way).

Note that because forced breaks are predictable, ranges on both sides of
forced breaks could in theory be fixed, however when the ranges cover
page spreads rather than single pages, this fix would only work for
"page-break: left", so I didn't bother.