[xproc-maven-plugin] Move parent to own module
The parent POM being the aggregator at the same time was causing trouble
for the super-build because aggregators are not mvn-installed.

Note that in the super-build, all poms including aggregators are
initially installed "manually" via the Makefile. While these manually
installed POMs are recognized by Maven when resolving the build
dependencies, they do not seem to be recognized by pax-exam-helper.