[dotify] Attempt to fix volume splitting failure
I commented out some code that is supposed to "fix a problem where the
volume overhead pushes the volume count up once the volume offset has
been set".

I'm not sure what exactly this problem was. I have tried to trace when
the code was added, but it dates back to the beginning of the Git
history (and haven't checked the SVN history). The hope is that
because several volume splitting enhancements have been made since
this code was added, it is not relevant anymore.