[mod-nlb] Test the hyphenation table
by checking the 571891 standard + 3556 non-standard input words.

For now the standard and non-standard hyphenations are processed
separately with the result that some bad hyphens can not be
detected. A possible improvement would be to process all data in a way
that all hyphens (standard and non-standard) for a certain word are
known.

The test currently fails. We could add a white list of bad and missed
hyphens that are acceptable.