Recompile hyphenation patterns with weighted words
From the ±824000 input words, only the ±225000 most frequent ones are
considered, and the words are weighted according to their
frequency. I was able to increase pat_finish from 7 to 9.

This results in ±187000 patterns (174000 standard and 13000
non-standard), 10 bad hyphens and 95727 missed. As expected, the
bad hyphens happen in words with low frequency.