EPUB: fix a volume breaking test
Note that this fixes the test but does not make EPUB behave exactly
like DTBook. It is difficult to achieve exactly the same behavior
because the document structure of EPUB is different from DTBook. As I
explained in https://github.com/nlbdev/pipeline/issues/119#issuecomment-382123073
the document structure is important for volume breaking.